Add 'enabled' property to MenuSub and MenuItem components (#531)
The 'enabled' property has been introduced into the MenuSub and MenuItem components. The render function in both components is now also checking this 'enabled' property. If a component is not enabled, it will now return an empty string and thus not render. This provides more flexibility when controlling the visibility of the components.