Add Pagination component and integrate into Table component (#523)
* Add Pagination component and integrate into Table component

A new Pagination component has been created and added to the MaryServiceProvider. This component has been integrated into the Table component to manage table pagination. The addition of Pagination provides greater control and customization options for page number displays. This update also reduces complexity within the Table component by offloading pagination responsibilities.

* Change default value of $perPage, update isShowable condition

The default value of $perPage was changed from an empty string to null in the Table component to avoid potential issues. Also, the condition in the isShowable method in the Pagination component was updated to check for a non-empty model name instead of a non-null one for better error handling.

* Add conditional check for pagination perPage parameter

Ensure that the pagination component only binds the live model if the perPage parameter is set. This avoids potential issues when the perPage value is missing.

---------

Co-authored-by: Robson Tenório <rrtenorio@gmail.com>