Editor/Markdow: allow to change contents on the fly (#572)
* Fixes: #564

Add refreshEditor method to Editor component

Introduce a new `refreshEditor` method in the Editor component and enhance the editor initialization with a watcher for content updates, ensuring synchronization between the editor content and the Livewire model. This update improves the component's reactivity and handling of dynamic content changes.

* Fixes: #564

Add refreshEditor method to Editor component

Introduce a new `refreshEditor` method in the Editor component and enhance the editor initialization with a watcher for content updates, ensuring synchronization between the editor content and the Livewire model. This update improves the component's reactivity and handling of dynamic content changes.

* Fixes: #564

Add refreshEditor method to Editor component

Introduce a new `refreshEditor` method in the Editor component and enhance the editor initialization with a watcher for content updates, ensuring synchronization between the editor content and the Livewire model. This update improves the component's reactivity and handling of dynamic content changes.

* WIP

---------

Co-authored-by: Robson Tenório <rrtenorio@gmail.com>