feat: remove live attribute
This makes the `wire:model.live` work seemlessly if the `mode` is set
`range`. This implementation checks if `mode` is set to `range` and the
`wire:model.live` attribute is present, then it rewrites the attribute
but setting it back to `wire:model` and adding a new attribute called
`live` which is used to check weather or not to attach the `x-on:change`
listener since we will removed the `wire:model.live` in the setup
function.