feat: make input inside a modal autofucosable
This commit makes an input inside a modal autofucosable by adding the
`focus_event` attribute to the modal and `modal_autofucos` attribute to
the input which should be focused. This has to be done as the autofocus
HTML attribute only sets focus for inputs that are visible on page load
which are not the case for most modal.