Add scroll and maxHeight properties to Dropdown
Occasionally I've come across dropdowns that i need to cut off to a certain height to make them look good.  This change adds a new property to Dropdown that adds an overflow and fixed height to allow the inner elements to be scrollable. 

Default behavior is the same.

Adding only 'scroll' will add the overflow and a default max height.  

Specifying scroll and max-height allows you to customize the height of the menu.